Put the drops back to floats

+++ EpicLoot/LootConfig.cs
@@ -24,9 +24,9 @@ namespace EpicLoot
     {
         public string Object;
         public string RefObject;
-        public int[][] Drops;
-        public int[][] Drops2;
-        public int[][] Drops3;
+        public float[][] Drops;
+        public float[][] Drops2;
+        public float[][] Drops3;
         public LootDrop[] Loot;
         public LootDrop[] Loot2;
         public LootDrop[] Loot3;
